pyproject.toml
[project]
name = "nicks-watch-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "pydantic-settings>=2.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "playwright>=1.44",
    "apscheduler>=3.10",
    "google-auth>=2.29",
    "google-api-python-client>=2.130",
    "anthropic>=0.30",
    "chrono24>=0.1",
    "mercapi>=0.1",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-httpx>=0.30",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
